Allow setting `use_logger` config to false
A change in configuration pattern matching meant that any value that did
not result in a `to_string` value of `"true"` would cause the Bugsnag
app to crash on launch. This was raised by @IceDragon200 in #39.

This resolves that issue by moving from a `case` to an `if` (since there
is only one case that requires action). There is also a test which
ensures a lack of explosion.

Closes #39.