Ensure tasks under Bugsnag.TaskSupervisor don't spill into other tests
A test in `BugsnagTest` was causing a task to be started by the
reporter, under `Bugsnag.TaskSupervisor`, but it wasn't waiting for the
task to finish or checking anything in it.

This was cauing tests in `Bugsnag.LoggerTest` to fail when the
problematic test was executed before them. The dangling task was getting
executed while the logger tests were running, causing `HTTPMock` to be
called twice which crashes the Mox process and every other test that
depend on it.

This commit adds a snippet that is executed after each test in
`BugsnagTest` to terminate possible dangling tasks in the supervisor
and so prevent this issue from ocurring again.