Add option to set error class instead of infering from the exception
This commit introduces a new option, `error_class`, which can be passed
to `Bugsnag.Reporter.report/2` and `Bugsnag.Reporter.sync_report/2` to
determine the type of the error being reported instead of determining
the error class using only the module of the exception struct.

The default behavior is still the same, if the option is not passed the
error class will be the module of the given exception struct (i.e.
`exception.__struct__`).